Deny unsupported SSH request types.
The built in SSH server only supports SSH request types of "env" and
"exec" while just ignoring others. This violates the SSH protocol, as
SSH clients might be waiting for an answer. This results in hanging SSH
connections, for instance when using a normal SSH shell client directly
and not git.

This change simply denies unsupprted SSH requests types such as
"pty-req" and "shell" by sending an appropiate reply to the client.

Signed-off-by: Dennis Keitzel <github@pinshot.net>